# apispec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/optimizing/result": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Optimizing Result",
        "operationId": "optimizing_result_optimizing_result_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingResult"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimizing/pt_to_onnx_fp32": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Pt To Onnx Fp32 Route",
        "operationId": "pt_to_onnx_fp32_route_optimizing_pt_to_onnx_fp32_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimizing/pt_to_onnx_fp16": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Pt To Onnx Fp16 Route",
        "operationId": "pt_to_onnx_fp16_route_optimizing_pt_to_onnx_fp16_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimizing/prune_unstructured": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Prune Unstructured Route",
        "operationId": "prune_unstructured_route_optimizing_prune_unstructured_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimizing/prune_structured": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Prune Structured Route",
        "operationId": "prune_structured_route_optimizing_prune_structured_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/optimizing/check_model_stats": {
      "post": {
        "tags": [
          "Optimizing"
        ],
        "summary": "Check Model Stats Route",
        "operationId": "check_model_stats_route_optimizing_check_model_stats_post",
        "parameters": [
          {
            "name": "uid",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "0001",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizingRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Action": {
        "type": "string",
        "enum": [
          "prune_unstructured",
          "prune_structured",
          "pt_to_onnx",
          "onnx_to_trt",
          "onnx_to_trt_int8",
          "check_model_stats"
        ],
        "title": "Action"
      },
...
      "CheckModelStatsParams": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "check_model_stats",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path"
        ],
        "title": "CheckModelStatsParams"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnnxToTrtInt8Params": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "onnx_to_trt_int8",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "output_path": {
            "type": "string",
            "title": "Output Path"
          },
          "calibrator": {
            "type": "string",
            "title": "Calibrator"
          },
          "device": {
            "type": "string",
            "enum": [
              "gpu",
              "dla",
              "cpu"
            ],
            "title": "Device",
            "default": "gpu"
          },
          "mixed_fp16": {
            "type": "boolean",
            "title": "Mixed Fp16",
            "default": true
          },
          "sparse": {
            "type": "boolean",
            "title": "Sparse",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path",
          "output_path",
          "calibrator"
        ],
        "title": "OnnxToTrtInt8Params"
      },
      "OnnxToTrtParams": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "onnx_to_trt",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "output_path": {
            "type": "string",
            "title": "Output Path"
          },
          "device": {
            "type": "string",
            "enum": [
              "gpu",
              "dla",
              "cpu"
            ],
            "title": "Device",
            "default": "gpu"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path",
          "output_path"
        ],
        "title": "OnnxToTrtParams"
      },
      "OptimizingInfo": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "oid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oid"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "workdir": {
            "type": "string",
            "title": "Workdir"
          }
        },
        "type": "object",
        "required": [
          "uid",
          "pid",
          "action",
          "workdir"
        ],
        "title": "OptimizingInfo"
      },
      "OptimizingRequest": {
        "properties": {
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "oid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oid"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "parameters": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PruneUnstructuredParams"
              },
              {
                "$ref": "#/components/schemas/PruneStructuredParams"
              },
              {
                "$ref": "#/components/schemas/PtToOnnxParams"
              },
              {
                "$ref": "#/components/schemas/OnnxToTrtParams"
              },
              {
                "$ref": "#/components/schemas/OnnxToTrtInt8Params"
              },
              {
                "$ref": "#/components/schemas/CheckModelStatsParams"
              }
            ],
            "title": "Parameters",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "check_model_stats": "#/components/schemas/CheckModelStatsParams",
                "onnx_to_trt": "#/components/schemas/OnnxToTrtParams",
                "onnx_to_trt_int8": "#/components/schemas/OnnxToTrtInt8Params",
                "prune_structured": "#/components/schemas/PruneStructuredParams",
                "prune_unstructured": "#/components/schemas/PruneUnstructuredParams",
                "pt_to_onnx": "#/components/schemas/PtToOnnxParams"
              }
            }
          },
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "pid",
          "action",
          "parameters"
        ],
        "title": "OptimizingRequest"
      },
      "OptimizingResult": {
        "properties": {
          "uid": {
            "type": "string",
            "title": "Uid"
          },
          "pid": {
            "type": "string",
            "title": "Pid"
          },
          "oid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Oid"
          },
          "action": {
            "$ref": "#/components/schemas/Action"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "artifacts_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts Path"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "uid",
          "pid",
          "action",
          "status"
        ],
        "title": "OptimizingResult"
      },
...
      "PruneStructuredParams": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "prune_structured",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "output_path": {
            "type": "string",
            "title": "Output Path"
          },
          "amount": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Amount",
            "default": 0.2
          },
          "pruning_type": {
            "type": "string",
            "const": "ln_structured",
            "title": "Pruning Type",
            "default": "ln_structured"
          },
          "dim": {
            "type": "integer",
            "title": "Dim",
            "default": 0
          },
          "n": {
            "type": "integer",
            "title": "N",
            "default": 2
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path",
          "output_path"
        ],
        "title": "PruneStructuredParams"
      },
      "PruneUnstructuredParams": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "prune_unstructured",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "output_path": {
            "type": "string",
            "title": "Output Path"
          },
          "amount": {
            "type": "number",
            "maximum": 1,
            "minimum": 0,
            "title": "Amount",
            "default": 0.2
          },
          "pruning_type": {
            "type": "string",
            "const": "l1_unstructured",
            "title": "Pruning Type",
            "default": "l1_unstructured"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path",
          "output_path"
        ],
        "title": "PruneUnstructuredParams"
      },
      "PtToOnnxParams": {
        "properties": {
          "info": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptimizingInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "type": "string",
            "const": "pt_to_onnx",
            "title": "Kind"
          },
          "input_path": {
            "type": "string",
            "title": "Input Path"
          },
          "output_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Path"
          },
          "input_size": {
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Input Size",
            "default": [640, 640]
          },
          "batch_size": {
            "type": "integer",
            "title": "Batch Size",
            "default": 1
          },
          "channels": {
            "type": "integer",
            "title": "Channels",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "kind",
          "input_path"
        ],
        "title": "PtToOnnxParams"
      },
...
      "Status": {
        "type": "string",
        "enum": [
          "started",
          "completed",
          "failed"
        ],
        "title": "Status"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}